+++ tests/encryption.rs
@@ -4,9 +4,10 @@
 //
 // Copyright (c) DUSK NETWORK. All rights reserved.
 
+use dusk_bytes::Serializable;
 use dusk_jubjub::{JubJubScalar, GENERATOR_EXTENDED};
 use ff::Field;
-use jubjub_elgamal::{decrypt, encrypt, DecryptionOrigin};
+use jubjub_elgamal::{DecryptFrom, Encryption};
 use rand::rngs::StdRng;
 use rand::SeedableRng;
 
@@ -21,23 +22,45 @@ fn encrypt_decrypt() {
 
     // Encrypt using a fresh random value 'blinder'
     let blinder = JubJubScalar::random(&mut rng);
-    let (c1, c2, shared_key) =
-        encrypt(&pk, &message, &GENERATOR_EXTENDED, &blinder);
+    let (ciphertext, shared_key) =
+        Encryption::encrypt(&pk, &message, None, &blinder);
 
     // Assert decryption using the secret key
-    let dec_message = decrypt(&DecryptionOrigin::FromSecretKey(sk), &(c1, c2));
+    let dec_message = ciphertext.decrypt(&DecryptFrom::SecretKey(sk));
     assert_eq!(message, dec_message);
 
     // Assert decryption using the shared key
-    let dec_message =
-        decrypt(&DecryptionOrigin::FromSharedKey(shared_key), &(c1, c2));
+    let dec_message = ciphertext.decrypt(&DecryptFrom::SharedKey(shared_key));
     assert_eq!(message, dec_message);
 
     // Assert decryption using an incorrect secret key
     let wrong_sk = JubJubScalar::random(&mut rng);
     let dec_message_wrong =
-        decrypt(&DecryptionOrigin::FromSecretKey(wrong_sk), &(c1, c2));
+        ciphertext.decrypt(&DecryptFrom::SecretKey(wrong_sk));
     assert_ne!(message, dec_message_wrong);
+
+    // encrypt / decrypt plaintext using custom generator
+    let custom_gen = GENERATOR_EXTENDED * JubJubScalar::random(&mut rng);
+    let custom_pk = custom_gen * sk;
+
+    let (custom_enc, _) =
+        Encryption::encrypt(&custom_pk, &message, Some(&custom_gen), &blinder);
+
+    let dec_message = custom_enc.decrypt(&DecryptFrom::SecretKey(sk));
+    assert_eq!(message, dec_message);
+}
+
+#[test]
+fn test_bytes() {
+    let mut rng = StdRng::seed_from_u64(0xc0b);
+    let point = GENERATOR_EXTENDED * &JubJubScalar::random(&mut rng);
+
+    let ciphertext = Encryption::new(point, point);
+
+    assert_eq!(
+        ciphertext,
+        Encryption::from_bytes(&ciphertext.to_bytes()).unwrap()
+    );
 }
 
 #[cfg(feature = "zk")]
@@ -47,11 +70,10 @@ mod zk {
     };
     use dusk_plonk::prelude::*;
     use ff::Field;
-    use jubjub_elgamal::encrypt;
     use jubjub_elgamal::zk::{
-        decrypt as decrypt_gadget, encrypt as encrypt_gadget,
-        DecryptionOrigin as DecryptionOriginZK,
+        DecryptFrom as DecryptFromZK, Encryption as EncryptionZK,
     };
+    use jubjub_elgamal::Encryption;
     use rand::rngs::StdRng;
     use rand::SeedableRng;
 
@@ -64,8 +86,7 @@ mod zk {
         secret_key: JubJubScalar,
         plaintext: JubJubAffine,
         r: JubJubScalar,
-        expected_ciphertext_1: JubJubAffine,
-        expected_ciphertext_2: JubJubAffine,
+        expected_ciphertext: Encryption,
     }
 
     impl<const MUST_PASS: bool> ElGamalCircuit<MUST_PASS> {
@@ -74,20 +95,14 @@ mod zk {
             secret_key: &JubJubScalar,
             plaintext: &JubJubExtended,
             r: &JubJubScalar,
-            expected_ciphertext_1: &JubJubExtended,
-            expected_ciphertext_2: &JubJubExtended,
+            expected_ciphertext: &Encryption,
         ) -> Self {
             Self {
                 public_key: JubJubAffine::from(public_key),
                 secret_key: *secret_key,
                 plaintext: JubJubAffine::from(plaintext),
                 r: *r,
-                expected_ciphertext_1: JubJubAffine::from(
-                    expected_ciphertext_1,
-                ),
-                expected_ciphertext_2: JubJubAffine::from(
-                    expected_ciphertext_2,
-                ),
+                expected_ciphertext: *expected_ciphertext,
             }
         }
     }
@@ -101,39 +116,32 @@ mod zk {
             let r = composer.append_witness(self.r);
 
             // encrypt plaintext using the public key
-            let (ciphertext_1, ciphertext_2, shared_key) =
-                encrypt_gadget(composer, public_key, plaintext, None, r)?;
+            let (ciphertext, shared_key) = EncryptionZK::encrypt(
+                composer, public_key, plaintext, None, r,
+            )?;
 
             // only for the 'encrypt_decrypt' test
             if MUST_PASS {
                 // assert that the ciphertext is as expected
                 composer.assert_equal_public_point(
-                    ciphertext_1,
-                    self.expected_ciphertext_1,
+                    *ciphertext.c1(),
+                    self.expected_ciphertext.c1(),
                 );
                 composer.assert_equal_public_point(
-                    ciphertext_2,
-                    self.expected_ciphertext_2,
+                    *ciphertext.c2(),
+                    self.expected_ciphertext.c2(),
                 );
 
                 // decrypt with sk
-                let dec_plaintext = decrypt_gadget(
-                    composer,
-                    &DecryptionOriginZK::FromSecretKey(secret_key),
-                    ciphertext_1,
-                    ciphertext_2,
-                );
+                let dec_plaintext = ciphertext
+                    .decrypt(composer, &DecryptFromZK::SecretKey(secret_key));
 
                 // assert decoded plaintext is the same as the original
                 composer.assert_equal_point(dec_plaintext, plaintext);
 
                 // decrypt with shared key
-                let dec_plaintext = decrypt_gadget(
-                    composer,
-                    &DecryptionOriginZK::FromSharedKey(shared_key),
-                    ciphertext_1,
-                    ciphertext_2,
-                );
+                let dec_plaintext = ciphertext
+                    .decrypt(composer, &DecryptFromZK::SharedKey(shared_key));
                 composer.assert_equal_point(dec_plaintext, plaintext);
 
                 // encrypt / decrypt plaintext using custom generator
@@ -142,7 +150,7 @@ mod zk {
                 );
                 let custom_pk =
                     composer.component_mul_point(secret_key, custom_gen);
-                let (custom_c1, custom_c2, _) = encrypt_gadget(
+                let (custom_enc, _) = EncryptionZK::encrypt(
                     composer,
                     custom_pk,
                     plaintext,
@@ -150,12 +158,8 @@ mod zk {
                     r,
                 )?;
 
-                let custom_dec_plaintext = decrypt_gadget(
-                    composer,
-                    &DecryptionOriginZK::FromSecretKey(secret_key),
-                    custom_c1,
-                    custom_c2,
-                );
+                let custom_dec_plaintext = custom_enc
+                    .decrypt(composer, &DecryptFromZK::SecretKey(secret_key));
                 composer.assert_equal_point(custom_dec_plaintext, plaintext);
             }
 
@@ -172,7 +176,7 @@ mod zk {
 
         let message = GENERATOR_EXTENDED * JubJubScalar::from(1234u64);
         let r = JubJubScalar::random(&mut rng);
-        let (c1, c2, _) = encrypt(&pk, &message, &GENERATOR_EXTENDED, &r);
+        let (ciphertext, _) = Encryption::encrypt(&pk, &message, None, &r);
 
         let pp = PublicParameters::setup(1 << CAPACITY, &mut rng).unwrap();
 
@@ -188,8 +192,7 @@ mod zk {
                     &sk,
                     &message,
                     &r,
-                    &c1.into(),
-                    &c2.into(),
+                    &ciphertext,
                 ),
             )
             .expect("failed to prove");
@@ -217,8 +220,7 @@ mod zk {
         let r = JubJubScalar::random(&mut rng);
 
         // not involved in this test
-        let c1 = message;
-        let c2 = message;
+        let ciphertext = Encryption::new(message, message);
 
         let pp = PublicParameters::setup(1 << CAPACITY, &mut rng).unwrap();
 
@@ -235,8 +237,7 @@ mod zk {
                     &sk,
                     &message,
                     &r,
-                    &c1.into(),
-                    &c2.into(),
+                    &ciphertext,
                 ),
             )
             .expect("failed to prove");